backend/requirements.txt
web3>=6.0.0
python-dotenv>=0.19.0
asyncio-mqtt>=0.11.0
coloredlogs>=15.0